+++ ktx/build.gradle
@@ -43,7 +43,7 @@ android {
 }
 
 dependencies {
-    api project(":library")
+    api project(':library')
     api Libs.AndroidX.core
 
     api Libs.Kotlin.stdlib
