+++ settings.gradle
@@ -19,5 +19,5 @@ include ':ktx'
 include ':dbx'
 include ':sample'
 include ':widgets'
-include ':testutils'
+include ':test-utils'
 include ':test-layouts'
