+++ test-utils/src/main/AndroidManifest.xml
